@@ -1,11 +1,11 @@
 name: kdtree
-version: 0.1.0
+version: 0.2.0
 description: A basic super fast Dart implementation of the k-dimensional (k-d) tree data structure.
 homepage: https://github.com/JankoLancer/kd-tree-dart
 
 environment:
-  sdk: '>=2.0.0 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
 
 dev_dependencies:
-  pedantic: ^1.8.0
-  test: ^1.6.0
+  pedantic: ^1.11.0
+  test: ^1.16.8
